[stable11] fix drone for 0.7

+++ autotest.sh
@@ -143,6 +143,10 @@ function cleanup_config {
 	if [ -f config/autotest-storage-swift.config.php ]; then
 		rm config/autotest-storage-swift.config.php
 	fi
+	# Remove autotest redis config
+	if [ -f config/redis.config.php ]; then
+		rm config/redis.config.php
+	fi
 }
 
 # restore config on exit
@@ -176,6 +180,10 @@ function execute_tests {
 	fi
 	cp tests/preseed-config.php config/config.php
 
+	if [ "$ENABLE_REDIS" == "true" ] ; then
+		cp tests/redis.config.php config/redis.config.php
+	fi
+
 	_DB=$DB
 
 	# drop database
@@ -200,7 +208,7 @@ function execute_tests {
                 fi
                 mysql -u "$DATABASEUSER" -powncloud -e "DROP DATABASE IF EXISTS $DATABASENAME" -h $DATABASEHOST || true
             else
-                DATABASEHOST=127.0.0.1
+                DATABASEHOST=mysql
             fi
 		fi
         echo "Waiting for MySQL initialisation ..."
@@ -234,7 +242,7 @@ function execute_tests {
 				fi
 				mysql -u "$DATABASEUSER" -powncloud -e "DROP DATABASE IF EXISTS $DATABASENAME" -h $DATABASEHOST || true
 			else
-				DATABASEHOST=127.0.0.1
+				DATABASEHOST=mysqlmb4
 			fi
 		fi
 
@@ -296,6 +304,17 @@ function execute_tests {
 
 			echo "Postgres is up."
 		else
+			if [ ! -z "$DRONE" ] ; then
+				DATABASEHOST=postgres
+			fi
+			echo "Waiting for Postgres to be available ..."
+			if ! apps/files_external/tests/env/wait-for-connection $DATABASEHOST 5432 60; then
+				echo "[ERROR] Waited 60 seconds, no response" >&2
+				exit 1
+			fi
+			echo "Give it 10 additional seconds ..."
+			sleep 10
+
 			if [ -z "$DRONE" ] ; then # no need to drop the DB when we are on CI
 				dropdb -U "$DATABASEUSER" "$DATABASENAME" || true
 			fi
